refactor(p2p): Split 6 oversized files into directory modules
Resolves #332 — split all Priority 1 and Priority 2 files identified
in the issue, bringing files >400 lines from 13 down to 6 (5 intentional
keep-as-is + 1 test file).

Splits performed:
- tracker.rs (903) → tracker/{mod,memory,operations,tests}.rs
- process.rs (769) → process/{mod,pushlog,bitswap,pending_dag}.rs
- event_handler.rs (761) → event_handler/{mod,gossip,pushlog,bitswap,doc_sync,branchable_sync}.rs
- handler.rs (622) → handler/{mod,inbound,pushlog,doc_sync,branchable_se}.rs
- p2p_host.rs (895) → p2p_host/{mod,swarm,two_stream,protocols}.rs
- command_handler.rs (596) → command_handler/{mod,network,messaging,pubsub,bitswap}.rs

Structural refactor only — no behavior changes. Public API unchanged.
