+++ crates/p2p/src/host/p2p_host/mod.rs
@@ -1,0 +1,309 @@
+//! P2P Host implementation for DefraDB.
+//!
+//! This module provides the main P2P host that manages the libp2p swarm,
+//! handles peer connections, and coordinates CRDT synchronization.
+
+mod protocols;
+mod swarm;
+mod two_stream;
+
+use std::collections::HashMap;
+use std::sync::Arc;
+use std::time::Duration;
+
+use futures::StreamExt;
+use iroh_bitswap::Store;
+use libp2p::{
+    identity::Keypair, noise, request_response, tcp, yamux, Multiaddr, PeerId, Swarm, SwarmBuilder,
+};
+use tokio::sync::mpsc;
+use tracing::{debug, info, warn};
+
+use crate::behaviour::DefraBehaviour;
+use crate::bitswap::ReplicatorRegistry;
+use crate::error::{Error, Result};
+use crate::message::PushLogReply;
+use crate::two_stream::{TwoStreamHandler, TwoStreamRunner};
+use crate::QueryId;
+
+use super::command::HostCommand;
+use super::event::HostEvent;
+use super::handle::P2PHostHandle;
+use super::ResponseChannel;
+
+/// Default idle connection timeout.
+const IDLE_CONNECTION_TIMEOUT: Duration = Duration::from_secs(60);
+
+/// P2P Host that manages the libp2p swarm.
+pub struct P2PHost<S: Store> {
+    pub(super) swarm: Swarm<DefraBehaviour<S>>,
+    pub(super) keypair: Keypair,
+    pub(super) command_rx: mpsc::Receiver<HostCommand>,
+    pub(super) event_tx: mpsc::Sender<HostEvent>,
+    pub(super) pending_requests: HashMap<
+        request_response::OutboundRequestId,
+        tokio::sync::oneshot::Sender<Result<PushLogReply>>,
+    >,
+    /// Replicator registry for access control
+    pub(super) replicators: Arc<ReplicatorRegistry>,
+    /// Two-stream handler for Go compatibility
+    pub(super) two_stream_handler: Arc<tokio::sync::Mutex<TwoStreamHandler>>,
+    /// Receiver for two-stream events
+    pub(super) two_stream_event_rx: mpsc::Receiver<crate::two_stream::TwoStreamEvent>,
+    /// Tracked spawned tasks for graceful shutdown
+    pub(super) spawned_tasks: tokio::task::JoinSet<()>,
+    /// Bitswap query abort handles for cancellation support
+    pub(super) bitswap_queries: HashMap<QueryId, tokio::task::AbortHandle>,
+    /// Per-peer addresses learned from connections and identify protocol.
+    /// Used by ActivePeers to return full multiaddrs (Go-compatible).
+    pub(super) peer_addrs: HashMap<PeerId, Multiaddr>,
+}
+
+impl<S: Store> P2PHost<S> {
+    /// Create a new P2P host with a generated identity and the given blockstore.
+    ///
+    /// # Arguments
+    ///
+    /// * `bitswap_store` - The blockstore for Bitswap block exchange
+    ///
+    /// # Returns
+    ///
+    /// A tuple of (P2PHost, P2PHostHandle, HostEvent receiver, ReplicatorRegistry).
+    /// The ReplicatorRegistry is shared and can be used for access control decisions.
+    pub async fn new(
+        bitswap_store: S,
+    ) -> Result<(
+        Self,
+        P2PHostHandle,
+        mpsc::Receiver<HostEvent>,
+        Arc<ReplicatorRegistry>,
+    )> {
+        let keypair = Keypair::generate_ed25519();
+        Self::with_keypair(keypair, bitswap_store).await
+    }
+
+    /// Create a new P2P host with pubsub optionally disabled.
+    pub async fn with_pubsub(
+        bitswap_store: S,
+        enable_pubsub: bool,
+    ) -> Result<(
+        Self,
+        P2PHostHandle,
+        mpsc::Receiver<HostEvent>,
+        Arc<ReplicatorRegistry>,
+    )> {
+        let keypair = Keypair::generate_ed25519();
+        Self::with_keypair_and_config(keypair, bitswap_store, enable_pubsub).await
+    }
+
+    /// Create a new P2P host with the given keypair and blockstore.
+    ///
+    /// # Arguments
+    ///
+    /// * `keypair` - The identity keypair for this node
+    /// * `bitswap_store` - The blockstore for Bitswap block exchange
+    ///
+    /// # Returns
+    ///
+    /// A tuple of (P2PHost, P2PHostHandle, HostEvent receiver, ReplicatorRegistry).
+    /// The ReplicatorRegistry is shared and can be used for access control decisions.
+    ///
+    /// # Note
+    ///
+    /// This must be called within a tokio runtime context as Bitswap spawns
+    /// background tasks for operations.
+    pub async fn with_keypair(
+        keypair: Keypair,
+        bitswap_store: S,
+    ) -> Result<(
+        Self,
+        P2PHostHandle,
+        mpsc::Receiver<HostEvent>,
+        Arc<ReplicatorRegistry>,
+    )> {
+        Self::with_keypair_and_config(keypair, bitswap_store, true).await
+    }
+
+    /// Create a new P2P host with the given keypair, blockstore, and pubsub config.
+    pub async fn with_keypair_and_config(
+        keypair: Keypair,
+        bitswap_store: S,
+        enable_pubsub: bool,
+    ) -> Result<(
+        Self,
+        P2PHostHandle,
+        mpsc::Receiver<HostEvent>,
+        Arc<ReplicatorRegistry>,
+    )> {
+        let local_peer_id = keypair.public().to_peer_id();
+        let local_public_key = keypair.public();
+
+        // Encode the public key as protobuf for use in P2P message metadata
+        let local_public_key_proto = local_public_key.encode_protobuf();
+
+        info!("Local peer ID: {}", local_peer_id);
+
+        // Pass keypair and blockstore to behaviour for message signing and block exchange
+        let behaviour = DefraBehaviour::new(
+            local_peer_id,
+            local_public_key,
+            keypair.clone(),
+            bitswap_store,
+            enable_pubsub,
+        )
+        .await
+        .map_err(|e| Error::Behaviour(e.to_string()))?;
+
+        // Enable TCP port reuse to match Go-libp2p behavior.
+        // Go-libp2p reuses the listen port for outgoing connections, so the
+        // remote side sees the listen address (not an ephemeral port).
+        // This is critical for ActivePeers to return correct addresses.
+        let tcp_config = tcp::Config::default().port_reuse(true);
+        let swarm = SwarmBuilder::with_existing_identity(keypair.clone())
+            .with_tokio()
+            .with_tcp(tcp_config, noise::Config::new, yamux::Config::default)
+            .map_err(|e| Error::Transport(e.to_string()))?
+            .with_behaviour(|_key| behaviour)
+            .expect("behaviour already constructed")
+            .with_swarm_config(|cfg| cfg.with_idle_connection_timeout(IDLE_CONNECTION_TIMEOUT))
+            .build();
+
+        let (command_tx, command_rx) = mpsc::channel(256);
+        let (event_tx, event_rx) = mpsc::channel(256);
+
+        let handle = P2PHostHandle::new(
+            command_tx,
+            local_public_key_proto,
+            local_peer_id,
+            keypair.clone(),
+        );
+
+        // Create the replicator registry for access control
+        let replicators = Arc::new(ReplicatorRegistry::new());
+
+        // Set up two-stream protocol for Go compatibility
+        let mut control = swarm.behaviour().stream.new_control();
+        let request_streams = control
+            .accept(TwoStreamHandler::request_protocol())
+            .map_err(|_| Error::Behaviour("Failed to register request protocol".into()))?;
+        let response_streams = control
+            .accept(TwoStreamHandler::response_protocol())
+            .map_err(|_| Error::Behaviour("Failed to register response protocol".into()))?;
+        let se_request_streams = control
+            .accept(TwoStreamHandler::se_request_protocol())
+            .map_err(|_| Error::Behaviour("Failed to register SE request protocol".into()))?;
+        let se_response_streams = control
+            .accept(TwoStreamHandler::se_response_protocol())
+            .map_err(|_| Error::Behaviour("Failed to register SE response protocol".into()))?;
+
+        let handler = TwoStreamHandler::new(control);
+        let pending = handler.pending_responses();
+        let two_stream_handler = Arc::new(tokio::sync::Mutex::new(handler));
+        let (two_stream_event_tx, two_stream_event_rx) = mpsc::channel(256);
+
+        // Spawn the two-stream runner as a background task
+        let runner = TwoStreamRunner::new(
+            pending,
+            request_streams,
+            response_streams,
+            se_request_streams,
+            se_response_streams,
+            two_stream_event_tx,
+        );
+        tokio::spawn(runner.run());
+
+        let host = Self {
+            swarm,
+            keypair,
+            command_rx,
+            event_tx,
+            pending_requests: HashMap::new(),
+            replicators: Arc::clone(&replicators),
+            two_stream_handler,
+            two_stream_event_rx,
+            spawned_tasks: tokio::task::JoinSet::new(),
+            bitswap_queries: HashMap::new(),
+            peer_addrs: HashMap::new(),
+        };
+
+        Ok((host, handle, event_rx, replicators))
+    }
+
+    /// Get the local peer ID.
+    pub fn local_peer_id(&self) -> PeerId {
+        *self.swarm.local_peer_id()
+    }
+
+    /// Get the keypair.
+    pub fn keypair(&self) -> &Keypair {
+        &self.keypair
+    }
+
+    /// Run the P2P host event loop.
+    ///
+    /// This method runs until shutdown is requested.
+    pub async fn run(mut self) {
+        loop {
+            // Biased select: process swarm events before commands.
+            // This ensures ConnectionEstablished events (which update peer_addrs)
+            // are processed before PeerAddresses commands read them.
+            tokio::select! {
+                biased;
+
+                event = self.swarm.select_next_some() => {
+                    self.handle_swarm_event(event).await;
+                }
+                // Handle two-stream protocol events (Go compatibility)
+                Some(two_stream_event) = self.two_stream_event_rx.recv() => {
+                    self.handle_two_stream_event(two_stream_event).await;
+                }
+                command = self.command_rx.recv() => {
+                    match command {
+                        Some(cmd) => {
+                            if !self.handle_command(cmd).await {
+                                break;
+                            }
+                        }
+                        None => {
+                            info!("Command channel closed, shutting down");
+                            break;
+                        }
+                    }
+                }
+            }
+        }
+
+        info!("P2P host shutdown complete");
+    }
+
+    /// Dial a peer at the given addresses.
+    pub(super) fn dial_peer(&mut self, peer_id: PeerId, addrs: Vec<Multiaddr>) -> Result<()> {
+        for addr in addrs {
+            let dial_addr = addr.with(libp2p::multiaddr::Protocol::P2p(peer_id));
+            match self.swarm.dial(dial_addr.clone()) {
+                Ok(_) => {
+                    debug!("Dialing peer {} at {}", peer_id, dial_addr);
+                    return Ok(());
+                }
+                Err(e) => {
+                    warn!("Failed to dial {}: {}", dial_addr, e);
+                }
+            }
+        }
+        Err(Error::Dial(format!("Failed to dial peer {}", peer_id)))
+    }
+
+    /// Send a PushLog response through the given channel.
+    pub fn send_pushlog_response(&mut self, channel: ResponseChannel, response: PushLogReply) {
+        if let Err(resp) = self
+            .swarm
+            .behaviour_mut()
+            .send_pushlog_response(channel.into_inner(), response)
+        {
+            tracing::error!(
+                "Failed to send PushLog response: message_id={}",
+                resp.message_id
+            );
+        }
+    }
+}
