+++ crates/p2p/src/sync/manager/process/bitswap.rs
@@ -1,0 +1,166 @@
+//! Bitswap query tracking and block storage.
+
+use cid::Cid;
+
+use blockstore::Blockstore;
+
+use crate::error::{Error, Result};
+use crate::sync::manager::events::SyncEvent;
+use crate::QueryId;
+
+use super::SyncManager;
+
+impl<B: Blockstore + 'static> SyncManager<B> {
+    /// Register a Bitswap query for tracking.
+    ///
+    /// This maps the QueryId to the root CID so we can identify
+    /// which DAG a completion event belongs to.
+    pub fn register_query(&self, query_id: QueryId, root_cid: Cid) {
+        self.query_to_root.write().insert(query_id, root_cid);
+    }
+
+    /// Handle Bitswap query completion.
+    ///
+    /// Called when a Bitswap sync completes (success or failure).
+    pub async fn handle_bitswap_complete(
+        &self,
+        query_id: QueryId,
+        success: bool,
+        error: Option<String>,
+    ) -> Result<()> {
+        // Find the root CID for this query
+        let root_cid = match self.query_to_root.write().remove(&query_id) {
+            Some(cid) => cid,
+            None => {
+                tracing::debug!(
+                    query_id = ?query_id,
+                    "Bitswap complete for unknown query, ignoring"
+                );
+                return Ok(());
+            }
+        };
+
+        if success {
+            // All blocks fetched - emit BlockReceived for the root
+            let dag = self.pending_dags.write().remove(&root_cid);
+            match dag {
+                Some(dag) => {
+                    tracing::info!(
+                        cid = %root_cid,
+                        doc_id = %dag.doc_id,
+                        "Bitswap sync complete, emitting BlockReceived"
+                    );
+
+                    if self
+                        .event_tx
+                        .send(SyncEvent::BlockReceived {
+                            cid: root_cid,
+                            doc_id: dag.doc_id,
+                            collection_id: dag.collection_id,
+                            creator: dag.creator,
+                        })
+                        .await
+                        .is_err()
+                    {
+                        tracing::error!(
+                            cid = %root_cid,
+                            "Failed to send BlockReceived after Bitswap complete - receiver dropped"
+                        );
+                        return Err(Error::ChannelSend);
+                    }
+                }
+                None => {
+                    // This can happen if the DAG was processed by another path,
+                    // cleaned up, or if there's a race condition
+                    tracing::warn!(
+                        cid = %root_cid,
+                        "Bitswap sync completed but no pending DAG found - \
+                         DAG may have been processed by another path or cleaned up"
+                    );
+                }
+            }
+        } else {
+            // Sync failed - emit error, clean up
+            self.pending_dags.write().remove(&root_cid);
+
+            let error_msg = error.unwrap_or_else(|| "Bitswap sync failed".to_string());
+            tracing::warn!(
+                cid = %root_cid,
+                error = %error_msg,
+                "Bitswap sync failed"
+            );
+
+            if self
+                .event_tx
+                .send(SyncEvent::SyncError {
+                    cid: root_cid,
+                    error: error_msg,
+                })
+                .await
+                .is_err()
+            {
+                tracing::warn!(
+                    cid = %root_cid,
+                    "Failed to send SyncError event - receiver dropped"
+                );
+                return Err(Error::ChannelSend);
+            }
+        }
+
+        Ok(())
+    }
+
+    /// Store a block received via Bitswap and check if pending DAGs can now proceed.
+    ///
+    /// This is called when blocks are fetched via Bitswap during DAG synchronization.
+    /// The block is stored in the blockstore, and we check if any pending DAGs are
+    /// now complete and can be processed.
+    ///
+    /// Returns `true` if the block was stored (not a duplicate).
+    pub async fn store_bitswap_block(&self, cid: &Cid, data: &[u8]) -> Result<bool> {
+        // Check if we already have the block
+        if self
+            .blockstore
+            .has(cid)
+            .await
+            .map_err(|e| Error::BlockstoreError(e.to_string()))?
+        {
+            tracing::debug!(
+                cid = %cid,
+                "Bitswap block already in blockstore (duplicate)"
+            );
+            return Ok(false);
+        }
+
+        // Store the block
+        if let Err(e) = self.blockstore.put(cid, data).await {
+            tracing::error!(
+                cid = %cid,
+                error = %e,
+                "Failed to store Bitswap block"
+            );
+            return Err(Error::BlockstoreError(e.to_string()));
+        }
+
+        tracing::info!(
+            cid = %cid,
+            data_len = data.len(),
+            "Stored Bitswap block in blockstore"
+        );
+
+        // Check if any pending DAGs can now proceed
+        // This is done by checking which pending DAGs were waiting for this CID
+        let pending = self.pending_dags.read().clone();
+        for (root_cid, pending_info) in pending {
+            if pending_info.missing.contains(cid) {
+                tracing::debug!(
+                    root_cid = %root_cid,
+                    received_cid = %cid,
+                    "Pending DAG received a missing block - will check completeness"
+                );
+            }
+        }
+
+        Ok(true)
+    }
+}
